Problem: /server/imageUpload.js is full of random methods that belong in /imports #468
Solution: Split methods into their respective contexts
